Move edr api to own file
